Split configs
