+++ package.json
@@ -37,12 +37,14 @@
     "@vue/test-utils": "^1.2.1",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "^27.0.5",
+    "hastscript": "^7.0.2",
     "imagemin-pngquant": "^9.0.2",
     "jest": "^27.2.3",
     "nuxt-content-git": "^2.0.9",
     "remark-code-blocks": "^2.0.1",
     "remark-code-extra": "^1.0.1",
     "remark-code-titles": "^0.1.2",
+    "remark-directive": "^2.0.1",
     "remark-prism": "^1.3.6",
     "sass": "^1.39.2",
     "sass-loader": "10",
